Dry run for resume script

+++ runner/runner.py
@@ -125,15 +125,13 @@ def main(
 
     start_run(start_nextflow_command, dry_run, skip_confirmation)
 
-    if dry_run:
-        logging.info(f"(dry) Writing resume.sh")
-    else:
-        write_resume_script(
-            results_dir,
-            config["settings"]["start_nextflow_analysis"],
-            out_csv,
-            stub_run,
-        )
+    write_resume_script(
+        results_dir,
+        config["settings"]["start_nextflow_analysis"],
+        out_csv,
+        stub_run,
+        dry_run,
+    )
 
     setup_results_links(config, results_dir, run_label, run_type, dry_run)
 
@@ -357,7 +355,9 @@ def start_run(
         LOG.info("(dry) " + joined_command)
 
 
-def write_resume_script(results_dir: Path, run_command: str, csv: Path, stub_run: bool):
+def write_resume_script(
+    results_dir: Path, run_command: str, csv: Path, stub_run: bool, dry_run: bool
+):
     resume_command_parts = [
         run_command,
         str(csv.absolute()),
@@ -367,7 +367,10 @@ def write_resume_script(results_dir: Path, run_command: str, csv: Path, stub_run
         resume_command_parts.append("'-stub-run'")
     resume_command = " ".join(resume_command_parts)
     resume_script = results_dir / "resume.sh"
-    resume_script.write_text(resume_command)
+    if not dry_run:
+        resume_script.write_text(resume_command)
+    else:
+        logging.info(f"(dry) Writing {resume_command} to {resume_script}")
 
 
 def setup_results_links(
